.history/requirements_20250623212828.txt
# Flask and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
basexclient==4.0.0

# XML and data processing
lxml==4.9.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1
